Refactoring structure functions
