+++ sakuracloud/resource_sakuracloud_disk.go
@@ -218,46 +218,3 @@ func setDiskResourceData(ctx context.Context, d *schema.ResourceData, client *AP
 	d.Set("zone", getZone(d, client))
 	return nil
 }
-
-func flattenDiskPlan(data *sacloud.Disk) string {
-	var plan string
-	switch data.DiskPlanID {
-	case types.DiskPlans.SSD:
-		plan = "ssd"
-	case types.DiskPlans.HDD:
-		plan = "hdd"
-	}
-	return plan
-}
-
-func expandDiskCreateRequest(d *schema.ResourceData) *sacloud.DiskCreateRequest {
-	var planID types.ID
-	switch d.Get("plan").(string) {
-	case "ssd":
-		planID = types.DiskPlans.SSD
-	case "hdd":
-		planID = types.DiskPlans.HDD
-	}
-
-	return &sacloud.DiskCreateRequest{
-		DiskPlanID:      planID,
-		Connection:      types.EDiskConnection(d.Get("connector").(string)),
-		SourceDiskID:    expandSakuraCloudID(d, "source_disk_id"),
-		SourceArchiveID: expandSakuraCloudID(d, "source_archive_id"),
-		SizeMB:          toSizeMB(d.Get("size").(int)),
-		Name:            d.Get("name").(string),
-		Description:     d.Get("description").(string),
-		Tags:            expandTags(d),
-		IconID:          expandSakuraCloudID(d, "icon_id"),
-	}
-}
-
-func expandDiskUpdateRequest(d *schema.ResourceData) *sacloud.DiskUpdateRequest {
-	return &sacloud.DiskUpdateRequest{
-		Connection:  types.EDiskConnection(d.Get("connector").(string)),
-		Name:        d.Get("name").(string),
-		Description: d.Get("description").(string),
-		Tags:        expandTags(d),
-		IconID:      expandSakuraCloudID(d, "icon_id"),
-	}
-}
