+++ sakuracloud/resource_sakuracloud_packet_filter.go
@@ -98,11 +98,7 @@ func resourceSakuraCloudPacketFilterCreate(d *schema.ResourceData, meta interfac
 	client, ctx, zone := getSacloudClient(d, meta)
 	pfOp := sacloud.NewPacketFilterOp(client)
 
-	pf, err := pfOp.Create(ctx, zone, &sacloud.PacketFilterCreateRequest{
-		Name:        d.Get("name").(string),
-		Description: d.Get("description").(string),
-		Expression:  expandPacketFilterExpressions(d),
-	})
+	pf, err := pfOp.Create(ctx, zone, expandPacketFilterCreateRequest(d))
 	if err != nil {
 		return fmt.Errorf("creating SakuraCloud PacketFilter is failed: %s", err)
 	}
@@ -170,76 +166,6 @@ func resourceSakuraCloudPacketFilterDelete(d *schema.ResourceData, meta interfac
 func setPacketFilterResourceData(ctx context.Context, d *schema.ResourceData, client *APIClient, data *sacloud.PacketFilter) error {
 	d.Set("name", data.Name)
 	d.Set("description", data.Description)
-	if err := d.Set("expressions", flattenPacketFilterExpressions(data)); err != nil {
-		return err
-	}
 	d.Set("zone", getZone(d, client))
-	return nil
-}
-
-func expandPacketFilterUpdateRequest(d *schema.ResourceData, pf *sacloud.PacketFilter) *sacloud.PacketFilterUpdateRequest {
-	expressions := pf.Expression
-	if d.HasChange("expressions") {
-		expressions = expandPacketFilterExpressions(d)
-	}
-
-	return &sacloud.PacketFilterUpdateRequest{
-		Name:        d.Get("name").(string),
-		Description: d.Get("description").(string),
-		Expression:  expressions,
-	}
-}
-
-func flattenPacketFilterExpressions(pf *sacloud.PacketFilter) []interface{} {
-	var expressions []interface{}
-	if len(pf.Expression) > 0 {
-		for _, exp := range pf.Expression {
-			expression := map[string]interface{}{}
-			protocol := exp.Protocol
-			switch protocol {
-			case types.Protocols.TCP, types.Protocols.UDP:
-				expression["source_network"] = exp.SourceNetwork
-				expression["source_port"] = exp.SourcePort
-				expression["destination_port"] = exp.DestinationPort
-			case types.Protocols.ICMP, types.Protocols.Fragment, types.Protocols.IP:
-				expression["source_network"] = exp.SourceNetwork
-			}
-			expression["protocol"] = exp.Protocol
-			expression["allow"] = exp.Action.IsAllow()
-			expression["description"] = exp.Description
-
-			expressions = append(expressions, expression)
-		}
-	}
-	return expressions
-}
-
-func expandPacketFilterExpressions(d resourceValueGettable) []*sacloud.PacketFilterExpression {
-	var expressions []*sacloud.PacketFilterExpression
-	for _, e := range d.Get("expressions").([]interface{}) {
-		expressions = append(expressions, expandPacketFilterExpression(&resourceMapValue{value: e.(map[string]interface{})}))
-	}
-	return expressions
-}
-
-func expandPacketFilterExpression(d resourceValueGettable) *sacloud.PacketFilterExpression {
-	action := "deny"
-	if d.Get("allow").(bool) {
-		action = "allow"
-	}
-
-	exp := &sacloud.PacketFilterExpression{
-		Protocol:      types.Protocol(d.Get("protocol").(string)),
-		SourceNetwork: types.PacketFilterNetwork(d.Get("source_network").(string)),
-		Action:        types.Action(action),
-		Description:   d.Get("description").(string),
-	}
-	if v, ok := d.GetOk("source_port"); ok {
-		exp.SourcePort = types.PacketFilterPort(v.(string))
-	}
-	if v, ok := d.GetOk("destination_port"); ok {
-		exp.DestinationPort = types.PacketFilterPort(v.(string))
-	}
-
-	return exp
+	return d.Set("expressions", flattenPacketFilterExpressions(data))
 }
