+++ sakuracloud/resource_sakuracloud_proxylb.go
@@ -278,21 +278,7 @@ func resourceSakuraCloudProxyLBCreate(d *schema.ResourceData, meta interface{}) 
 	client, ctx, _ := getSacloudClient(d, meta)
 	proxyLBOp := sacloud.NewProxyLBOp(client)
 
-	proxyLB, err := proxyLBOp.Create(ctx, &sacloud.ProxyLBCreateRequest{
-		Plan:           types.EProxyLBPlan(d.Get("plan").(int)),
-		HealthCheck:    expandProxyLBHealthCheck(d),
-		SorryServer:    expandProxyLBSorryServer(d),
-		BindPorts:      expandProxyLBBindPorts(d),
-		Servers:        expandProxyLBServers(d),
-		StickySession:  expandProxyLBStickySession(d),
-		Timeout:        expandProxyLBTimeout(d),
-		UseVIPFailover: d.Get("vip_failover").(bool),
-		Region:         types.EProxyLBRegion(d.Get("region").(string)),
-		Name:           d.Get("name").(string),
-		Description:    d.Get("description").(string),
-		Tags:           expandTags(d),
-		IconID:         expandSakuraCloudID(d, "icon_id"),
-	})
+	proxyLB, err := proxyLBOp.Create(ctx, expandProxyLBCreateRequest(d))
 	if err != nil {
 		return fmt.Errorf("creating SakuraCloud ProxyLB is failed: %s", err)
 	}
@@ -342,18 +328,7 @@ func resourceSakuraCloudProxyLBUpdate(d *schema.ResourceData, meta interface{}) 
 		return fmt.Errorf("could not read SakuraCloud ProxyLB[%s]: %s", d.Id(), err)
 	}
 
-	proxyLB, err = proxyLBOp.Update(ctx, proxyLB.ID, &sacloud.ProxyLBUpdateRequest{
-		HealthCheck:   expandProxyLBHealthCheck(d),
-		SorryServer:   expandProxyLBSorryServer(d),
-		BindPorts:     expandProxyLBBindPorts(d),
-		Servers:       expandProxyLBServers(d),
-		StickySession: expandProxyLBStickySession(d),
-		Timeout:       expandProxyLBTimeout(d),
-		Name:          d.Get("name").(string),
-		Description:   d.Get("description").(string),
-		Tags:          expandTags(d),
-		IconID:        expandSakuraCloudID(d, "icon_id"),
-	})
+	proxyLB, err = proxyLBOp.Update(ctx, proxyLB.ID, expandProxyLBUpdateRequest(d))
 	if err != nil {
 		fmt.Errorf("updating SakuraCloud ProxyLB[%s] is failed: %s", d.Id(), err)
 	}
@@ -450,225 +425,6 @@ func setProxyLBResourceData(ctx context.Context, d *schema.ResourceData, client 
 	}
 	if err := d.Set("certificate", flattenProxyLBCerts(certs)); err != nil {
 		return err
-	}
-	return nil
-}
-
-func flattenProxyLBBindPorts(proxyLB *sacloud.ProxyLB) []interface{} {
-	var bindPorts []interface{}
-	for _, bindPort := range proxyLB.BindPorts {
-		var headers []interface{}
-		for _, header := range bindPort.AddResponseHeader {
-			headers = append(headers, map[string]interface{}{
-				"header": header.Header,
-				"value":  header.Value,
-			})
-		}
-
-		bindPorts = append(bindPorts, map[string]interface{}{
-			"proxy_mode":        bindPort.ProxyMode,
-			"port":              bindPort.Port,
-			"redirect_to_https": bindPort.RedirectToHTTPS,
-			"support_http2":     bindPort.SupportHTTP2,
-			"response_header":   headers,
-		})
-	}
-	return bindPorts
-}
-
-func flattenProxyLBHealthCheck(proxyLB *sacloud.ProxyLB) []interface{} {
-	var results []interface{}
-	if proxyLB.HealthCheck != nil {
-		results = []interface{}{
-			map[string]interface{}{
-				"protocol":    proxyLB.HealthCheck.Protocol,
-				"delay_loop":  proxyLB.HealthCheck.DelayLoop,
-				"host_header": proxyLB.HealthCheck.Host,
-				"path":        proxyLB.HealthCheck.Path,
-			},
-		}
-	}
-	return results
-}
-
-func flattenProxyLBSorryServer(proxyLB *sacloud.ProxyLB) []interface{} {
-	var results []interface{}
-	if proxyLB.SorryServer != nil && proxyLB.SorryServer.IPAddress != "" {
-		results = []interface{}{
-			map[string]interface{}{
-				"ipaddress": proxyLB.SorryServer.IPAddress,
-				"port":      proxyLB.SorryServer.Port,
-			},
-		}
-	}
-	return results
-}
-
-func flattenProxyLBServers(proxyLB *sacloud.ProxyLB) []interface{} {
-	var results []interface{}
-	for _, server := range proxyLB.Servers {
-		results = append(results, map[string]interface{}{
-			"ipaddress": server.IPAddress,
-			"port":      server.Port,
-			"enabled":   server.Enabled,
-		})
-	}
-	return results
-}
-
-func flattenProxyLBCerts(certs *sacloud.ProxyLBCertificates) []interface{} {
-	if certs == nil {
-		return nil
-	}
-	proxylbCert := map[string]interface{}{
-		"server_cert":       certs.ServerCertificate,
-		"intermediate_cert": certs.IntermediateCertificate,
-		"private_key":       certs.PrivateKey,
-	}
-	if len(certs.AdditionalCerts) > 0 {
-		var additionalCerts []interface{}
-		for _, cert := range certs.AdditionalCerts {
-			additionalCerts = append(additionalCerts, map[string]interface{}{
-				"server_cert":       cert.ServerCertificate,
-				"intermediate_cert": cert.IntermediateCertificate,
-				"private_key":       cert.PrivateKey,
-			})
-		}
-		proxylbCert["additional_certificates"] = additionalCerts
-	}
-	return []interface{}{proxylbCert}
-}
-
-func flattenProxyLBStickySession(proxyLB *sacloud.ProxyLB) bool {
-	if proxyLB.StickySession != nil {
-		return proxyLB.StickySession.Enabled
-	}
-	return false
-}
-
-func flattenProxyLBTimeout(proxyLB *sacloud.ProxyLB) int {
-	if proxyLB.Timeout != nil {
-		return proxyLB.Timeout.InactiveSec
-	}
-	return 0
-}
-
-func expandProxyLBStickySession(d resourceValueGettable) *sacloud.ProxyLBStickySession {
-	stickySession := d.Get("sticky_session").(bool)
-	if stickySession {
-		return &sacloud.ProxyLBStickySession{
-			Enabled: true,
-			Method:  "cookie",
-		}
-	}
-	return nil
-}
-
-func expandProxyLBBindPorts(d resourceValueGettable) []*sacloud.ProxyLBBindPort {
-	var results []*sacloud.ProxyLBBindPort
-	if bindPorts, ok := getListFromResource(d, "bind_ports"); ok {
-		for _, bindPort := range bindPorts {
-			values := mapToResourceData(bindPort.(map[string]interface{}))
-			var headers []*sacloud.ProxyLBResponseHeader
-			if rawHeaders, ok := values.GetOk("response_header"); ok {
-				for _, rawHeader := range rawHeaders.([]interface{}) {
-					if rawHeader == nil {
-						continue
-					}
-					v := rawHeader.(map[string]interface{})
-					headers = append(headers, &sacloud.ProxyLBResponseHeader{
-						Header: v["header"].(string),
-						Value:  v["value"].(string),
-					})
-				}
-			}
-
-			results = append(results, &sacloud.ProxyLBBindPort{
-				ProxyMode:         types.EProxyLBProxyMode(values.Get("proxy_mode").(string)),
-				Port:              values.Get("port").(int),
-				RedirectToHTTPS:   values.Get("redirect_to_https").(bool),
-				SupportHTTP2:      values.Get("support_http2").(bool),
-				AddResponseHeader: headers,
-			})
-		}
-	}
-	return results
-}
-
-func expandProxyLBHealthCheck(d resourceValueGettable) *sacloud.ProxyLBHealthCheck {
-	if healthChecks, ok := getListFromResource(d, "health_check"); ok {
-		v := mapToResourceData(healthChecks[0].(map[string]interface{}))
-		protocol := v.Get("protocol").(string)
-		switch protocol {
-		case "http":
-			return &sacloud.ProxyLBHealthCheck{
-				Protocol:  types.ProxyLBProtocols.HTTP,
-				Path:      v.Get("path").(string),
-				Host:      v.Get("host_header").(string),
-				DelayLoop: v.Get("delay_loop").(int),
-			}
-		case "tcp":
-			return &sacloud.ProxyLBHealthCheck{
-				Protocol:  types.ProxyLBProtocols.TCP,
-				DelayLoop: v.Get("delay_loop").(int),
-			}
-		}
-	}
-	return nil
-}
-
-func expandProxyLBSorryServer(d resourceValueGettable) *sacloud.ProxyLBSorryServer {
-	if sorryServers, ok := getListFromResource(d, "sorry_server"); ok && len(sorryServers) > 0 {
-		v := mapToResourceData(sorryServers[0].(map[string]interface{}))
-		return &sacloud.ProxyLBSorryServer{
-			IPAddress: v.Get("ipaddress").(string),
-			Port:      v.Get("port").(int),
-		}
-	}
-	return nil
-}
-
-func expandProxyLBServers(d resourceValueGettable) []*sacloud.ProxyLBServer {
-	var results []*sacloud.ProxyLBServer
-	if servers, ok := getListFromResource(d, "servers"); ok && len(servers) > 0 {
-		for _, server := range servers {
-			v := mapToResourceData(server.(map[string]interface{}))
-			results = append(results, &sacloud.ProxyLBServer{
-				IPAddress: v.Get("ipaddress").(string),
-				Port:      v.Get("port").(int),
-				Enabled:   v.Get("enabled").(bool),
-			})
-		}
-	}
-	return results
-}
-
-func expandProxyLBTimeout(d resourceValueGettable) *sacloud.ProxyLBTimeout {
-	return &sacloud.ProxyLBTimeout{InactiveSec: d.Get("timeout").(int)}
-}
-
-func expandProxyLBCerts(d resourceValueGettable) *sacloud.ProxyLBCertificates {
-	// set cert
-	if certs, ok := getListFromResource(d, "certificate"); ok && len(certs) > 0 {
-		values := mapToResourceData(certs[0].(map[string]interface{}))
-		cert := &sacloud.ProxyLBCertificates{
-			ServerCertificate:       values.Get("server_cert").(string),
-			IntermediateCertificate: values.Get("intermediate_cert").(string),
-			PrivateKey:              values.Get("private_key").(string),
-		}
-
-		if rawAdditionalCerts, ok := getListFromResource(values, "additional_certificates"); ok && len(rawAdditionalCerts) > 0 {
-			for _, rawCert := range rawAdditionalCerts {
-				values := mapToResourceData(rawCert.(map[string]interface{}))
-				cert.AdditionalCerts = append(cert.AdditionalCerts, &sacloud.ProxyLBAdditionalCert{
-					ServerCertificate:       values.Get("server_cert").(string),
-					IntermediateCertificate: values.Get("intermediate_cert").(string),
-					PrivateKey:              values.Get("private_key").(string),
-				})
-			}
-		}
-
-		return cert
 	}
 	return nil
 }
