+++ sakuracloud/structure_note.go
@@ -1,0 +1,40 @@
+// Copyright 2016-2019 terraform-provider-sakuracloud authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sakuracloud
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/sacloud/libsacloud/v2/sacloud"
+)
+
+func expandNoteCreateRequest(d *schema.ResourceData) *sacloud.NoteCreateRequest {
+	return &sacloud.NoteCreateRequest{
+		Name:    d.Get("name").(string),
+		Tags:    expandTags(d),
+		IconID:  expandSakuraCloudID(d, "icon_id"),
+		Class:   d.Get("class").(string),
+		Content: d.Get("content").(string),
+	}
+}
+
+func expandNoteUpdateRequest(d *schema.ResourceData) *sacloud.NoteUpdateRequest {
+	return &sacloud.NoteUpdateRequest{
+		Name:    d.Get("name").(string),
+		Tags:    expandTags(d),
+		IconID:  expandSakuraCloudID(d, "icon_id"),
+		Class:   d.Get("class").(string),
+		Content: d.Get("content").(string),
+	}
+}
