+++ sakuracloud/structure_vpc_router.go
@@ -21,8 +21,27 @@ import (
 
 	"github.com/sacloud/libsacloud/v2/sacloud"
 	"github.com/sacloud/libsacloud/v2/sacloud/types"
+	"github.com/sacloud/libsacloud/v2/utils/builder"
 	"github.com/sacloud/libsacloud/v2/utils/builder/vpcrouter"
 )
+
+func expandVPCRouterBuilder(d resourceValueGettable, client *APIClient) *vpcrouter.Builder {
+	return &vpcrouter.Builder{
+		Name:                  d.Get("name").(string),
+		Description:           d.Get("description").(string),
+		Tags:                  expandTags(d),
+		IconID:                expandSakuraCloudID(d, "icon_id"),
+		PlanID:                expandVPCRouterPlanID(d),
+		NICSetting:            expandVPCRouterNICSetting(d),
+		AdditionalNICSettings: expandVPCRouterAdditionalNICSettings(d),
+		RouterSetting:         expandVPCRouterSettings(d),
+		SetupOptions: &builder.RetryableSetupParameter{
+			BootAfterBuild:        true,
+			NICUpdateWaitDuration: builder.DefaultNICUpdateWaitDuration,
+		},
+		Client: sacloud.NewVPCRouterOp(client),
+	}
+}
 
 func expandVPCRouterPlanID(d resourceValueGettable) types.ID {
 	plan := d.Get("plan").(string)
